+++ Poseidon/src/main/java/com/poseidon/service/TradeService.java
@@ -3,18 +3,23 @@ package com.poseidon.service;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import javax.transaction.Transactional;
+
 import org.mapstruct.factory.Mappers;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.poseidon.dto.TradeDto;
 import com.poseidon.entity.Trade;
+import com.poseidon.exception.ResourceNotFoundException;
 import com.poseidon.mapper.TradeMapper;
 import com.poseidon.repository.TradeRepository;
 
 @Service
+@Transactional(rollbackOn = Exception.class)
 public class TradeService {
 
+	@Autowired
 	private TradeMapper tradeMapper ;
 	
 	@Autowired
@@ -33,16 +38,20 @@ public class TradeService {
 		return tradeToCreate;
 	}
 	
-	public TradeDto read(Integer id) {
-		return tradeMapper.tradeToTradeDto(tradeRepository.findById(id).orElse(null));
+	public TradeDto read(Integer id) throws ResourceNotFoundException {
+		return tradeMapper.tradeToTradeDto(tradeRepository.findById(id).orElseThrow(()->new ResourceNotFoundException("Trade not found")));
 	}
 	
-	public TradeDto update(TradeDto tradeToUpdate) {
+	public TradeDto update(TradeDto tradeToUpdate) throws ResourceNotFoundException {
+		if(!tradeRepository.existsById(tradeToUpdate.getTradeId()))
+			throw new ResourceNotFoundException("Trade not found");
 		tradeRepository.save(tradeMapper.tradeDtoToTrade(tradeToUpdate));
 		return tradeToUpdate;
 	}
 	
-	public TradeDto delete(TradeDto tradeToDelete) {
+	public TradeDto delete(TradeDto tradeToDelete) throws ResourceNotFoundException {
+		if(!tradeRepository.existsById(tradeToDelete.getTradeId()))
+			throw new ResourceNotFoundException("Trade not found");
 		tradeRepository.delete(tradeMapper.tradeDtoToTrade(tradeToDelete));
 		return tradeToDelete;
 	}
