+++ Poseidon/src/main/java/com/poseidon/service/UserService.java
@@ -3,18 +3,24 @@ package com.poseidon.service;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import javax.transaction.Transactional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import com.poseidon.dto.UserDto;
 import com.poseidon.entity.User;
+import com.poseidon.exception.DuplicateUserException;
+import com.poseidon.exception.UserNotFoundException;
 import com.poseidon.mapper.UserMapper;
 import com.poseidon.repository.UserRepository;
 
 @Service
+@Transactional(rollbackOn = Exception.class)
 public class UserService {
 	
+	@Autowired
 	private UserMapper userMapper ;
 
 	@Autowired
@@ -31,23 +37,29 @@ public class UserService {
 		return usersDto;
 	}
 	
-	public UserDto create(UserDto userToCreate) {
+	public UserDto create(UserDto userToCreate) throws DuplicateUserException {
+		if(userRepository.findByUserName(userToCreate.getUserName()) != null)
+			throw new DuplicateUserException("A user with user name " + userToCreate.getUserName() + " is alreday registered");
 		userToCreate.setUserPassword(bCryptPasswordEncoder.encode(userToCreate.getUserPassword()));
 		userRepository.save(userMapper.userDtoToUser(userToCreate));
 		return userToCreate ;
 	}
 	
-	public UserDto read(Integer id) {
-		return userMapper.userToUserDto(userRepository.findById(id).orElse(null));
+	public UserDto read(Integer id) throws UserNotFoundException {
+		return userMapper.userToUserDto(userRepository.findById(id).orElseThrow(()->new UserNotFoundException("User not found")));
 	}
 	
-	public UserDto update(UserDto userToUpdate) {
+	public UserDto update(UserDto userToUpdate) throws UserNotFoundException {
+		if(!userRepository.existsById(userToUpdate.getUserId()))
+			throw new UserNotFoundException("User not found");
 		userToUpdate.setUserPassword(bCryptPasswordEncoder.encode(userToUpdate.getUserPassword()));
 		userRepository.save(userMapper.userDtoToUser(userToUpdate));
 		return userToUpdate ;
 	}
 	
-	public UserDto delete(UserDto userToDelete) {
+	public UserDto delete(UserDto userToDelete) throws UserNotFoundException {
+		if(!userRepository.existsById(userToDelete.getUserId()))
+			throw new UserNotFoundException("User not found");
 		userRepository.delete(userMapper.userDtoToUser(userToDelete));
 		return userToDelete;
 	}
